Move testcases of readonly array methods to a document_test.ts
